+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/ReadConsistencyStrategy.java
@@ -22,7 +22,8 @@ import java.util.Map;
  * in RequestOptions, CosmosClient or the default consistency level for an account unless
  * ReadConsistencyStrategy `DEFAULT` is used.
  * <p>
- * NOTE: The ReadConsistencyStrategy is currently only working when using direct mode
+ * NOTE: The ReadConsistencyStrategy is supported in direct mode, thin client (Gateway V2) mode,
+ * and gateway mode
  */
 @Beta(value = Beta.SinceVersion.V4_69_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
 public enum ReadConsistencyStrategy {